+++ Sources/Controllers/ALKBaseViewController.swift
@@ -29,6 +29,9 @@ open class ALKBaseViewController: UIViewController, ALKConfigurable {
         super.viewWillAppear(animated)
         self.navigationController?.navigationBar.barTintColor = configuration.navigationBarBackgroundColor
         self.navigationController?.navigationBar.tintColor = configuration.navigationBarItemColor
+
+        self.navigationController?.navigationBar.titleTextAttributes = [NSAttributedStringKey.foregroundColor:configuration.navigationBarTitleColor]
+
         self.navigationController?.navigationBar.isTranslucent = false
         if self.navigationController?.viewControllers.first != self {
             var backImage = UIImage.init(named: "icon_back", in: Bundle.applozic, compatibleWith: nil)