+++ Sources/Controllers/ALKConversationViewController.swift
@@ -74,11 +74,10 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
         }
     }
 
-    let tableView : UITableView = {
+      var tableView : UITableView = {
         let tv = UITableView(frame: .zero, style: .grouped)
         tv.separatorStyle   = .none
         tv.allowsSelection  = false
-        tv.backgroundColor  = UIColor.clear
         tv.clipsToBounds    = true
         tv.keyboardDismissMode = UIScrollViewKeyboardDismissMode.onDrag
         tv.accessibilityIdentifier = "InnerChatScreenTableView"
@@ -87,7 +86,6 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
 
     fileprivate let titleButton : UIButton = {
         let titleButton = UIButton(frame: CGRect(x: 0, y: 0, width: 200, height: 200))
-        titleButton.setTitleColor(UIColor.black, for: .normal)
         titleButton.titleLabel?.font  = UIFont.boldSystemFont(ofSize: 17.0)
         return titleButton
     }()
@@ -479,6 +477,7 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
     private func setupNavigation() {
 
         titleButton.setTitle(self.title, for: .normal)
+        titleButton.setTitleColor(self.configuration.navigationBarTitleColor, for: .normal)
         titleButton.addTarget(self, action: #selector(showParticipantListChat), for: .touchUpInside)
         titleButton.isEnabled = isGroupDetailActionEnabled
         self.navigationItem.titleView = titleButton